feat(UH2004): Update button functionality added
