+++ package.json
@@ -15,6 +15,7 @@
     "@mui/material": "^6.1.6",
     "@mui/x-data-grid": "^7.22.1",
     "@tanstack/react-query": "^5.59.16",
+    "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
@@ -23,6 +24,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
+    "@types/axios": "^0.14.4",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
