+++ src/hooks/useGetCustomers.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { fetchCustomers } from "../services/getCustomers";
-import { Customer } from "../types/listCustomers";
+import { Customer } from "../types/typesCustomers";
 
 export const CustomerHook = () => {
   return useQuery<Customer[]>({
