--- src/types/listCustomers.ts
@@ -1,13 +1,0 @@
-
-export type Customer = {
-    id: number;
-    nit: string;
-    name: string;
-    address: string;
-    city: string;
-    country: string;
-    phone: string;
-    email: string;
-    active: boolean;
-  };
-
